refactor: reorganize auth interfaces and move DTOs to separate package
