+++ core/ports/auth_client.go
@@ -3,8 +3,8 @@ package ports
 import (
 	"context"
 
-	"github.com/Nerzal/gocloak/v13"
 	"github.com/EstebanGitPro/motogo-backend/core/domain"
+	"github.com/Nerzal/gocloak/v13"
 )
 
 // KeycloakClient define las operaciones con Keycloak como servicio externo.
