+++ core/ports/repository.go
@@ -3,9 +3,9 @@ package ports
 import "github.com/EstebanGitPro/motogo-backend/core/domain"
 
 type Repository interface {
-	Save(person domain.Person) error
+	SavePerson(person domain.Person) error
 	GetPersonByEmail(email string) (*domain.Person, error)
 	GetPersonByID(id string) (*domain.Person, error)
-	Update(person domain.Person) error
-	Delete(id string) error
+	UpdatePerson(person domain.Person) error
+	DeletePerson(id string) error
 }
