+++ core/ports/service.go
@@ -2,17 +2,13 @@ package ports
 
 import (
 	"github.com/EstebanGitPro/motogo-backend/core/domain"
+	"github.com/EstebanGitPro/motogo-backend/core/dto"
 	"github.com/Nerzal/gocloak/v13"
 )
 
-// RegistrationResult contiene el resultado del registro con token de Keycloak
-type RegistrationResult struct {
-	Person domain.Person
-	Token  *gocloak.JWT
-}
 
 type Service interface {
-	RegisterPerson(person domain.Person) (*RegistrationResult, error)
-	Login(email, password string) (*gocloak.JWT, error)
+	RegisterPerson(person domain.Person) (*dto.RegistrationResult, error)
+	LoginPerson(email, password string) (*gocloak.JWT, error)
 	GetPersonByEmail(email string) (*domain.Person, error)
 }
