+++ handlers/person.go
@@ -28,6 +28,7 @@ type PersonResponse struct {
 	EmailVerified       bool   `json:"email_verified"`
 	PhoneNumberVerified bool   `json:"phone_number_verified"`
 	Role                string `json:"role"`
+	KeycloakUserID      string `json:"keycloak_user_id"`
 }
 
 // RegistrationResponse incluye los datos del usuario y el token JWT de Keycloak
@@ -43,6 +44,7 @@ type RegistrationResponse struct {
 type LoginRequest struct {
 	Email    string `json:"email" binding:"required,email"`
 	Password string `json:"password" binding:"required"`
+	
 }
 
 // LoginResponse devuelve el token JWT de Keycloak
